+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## CHANGELOG
 
-1. added fetch optimisation for imgOnly mode.
+1. added option to resize columnWidth for react-stack-grid in imgOnly mode.
+2. header display name fixed.